feat: migrate to chrome extension v2
