--- electron-builder.yml
@@ -1,33 +1,0 @@
-appId: com.takaaizu.vrchat-scheduler
-productName: VRChat Group Scheduler
-copyright: Copyright TakaAizu © 2026
-directories:
-  output: dist
-  buildResources: resources
-files:
-  - from: .
-    filter:
-      - package.json
-      - app
-publish:
-  provider: github
-  owner: takacore
-  repo: vrchat-group-scheduler
-
-mac:
-  target: dmg
-  icon: resources/icon.icns
-  hardenedRuntime: true
-  gatekeeperAssess: false
-  artifactName: "VRChatGroupScheduler-${version}-mac.${ext}"
-
-win:
-  target: nsis
-  icon: resources/icon.ico
-  artifactName: "VRChatGroupScheduler-Setup-${version}.${ext}"
-
-nsis:
-  oneClick: true
-  perMachine: false
-  allowToChangeInstallationDirectory: false
-  deleteAppDataOnUninstall: false
