--- main/ipc-handlers.js
@@ -1,95 +1,0 @@
-import { app, ipcMain } from 'electron';
-import { login, logout, verify2FA, getCurrentUser, getUserGroups, refreshUserGroups } from './vrchat.js';
-import { addPost, deletePost, getPosts } from './scheduler.js';
-import { checkForUpdates, getUpdateSettings, saveUpdateSettings, openDownloadPage, downloadUpdate, installUpdate } from './updater.js';
-
-export function registerIpcHandlers() {
-    // App
-    ipcMain.handle('app:get-version', () => {
-        return app.getVersion();
-    });
-
-    // Auth
-    ipcMain.handle('auth:login', async (_, { username, password }) => {
-        return await login(username, password);
-    });
-
-    ipcMain.handle('auth:verify-2fa', async (_, { code }) => {
-        return await verify2FA(code);
-    });
-
-    ipcMain.handle('auth:get-user', async () => {
-        return await getCurrentUser();
-    });
-
-    ipcMain.handle('auth:logout', async () => {
-        return await logout();
-    });
-
-    // Groups (filtered by permission in getUserGroups, with persistent cache)
-    ipcMain.handle('groups:get-all', async (_, { userId }) => {
-        if (!userId) {
-            const user = await getCurrentUser();
-            if (!user) throw new Error('Not logged in');
-            userId = user.id;
-        }
-        return await getUserGroups(userId);
-    });
-
-    ipcMain.handle('groups:refresh', async (event, { userId }) => {
-        if (!userId) {
-            const user = await getCurrentUser();
-            if (!user) throw new Error('Not logged in');
-            userId = user.id;
-        }
-        const onProgress = (progress) => {
-            try {
-                event.sender.send('groups:scan-progress', progress);
-            } catch (e) {
-                // Window may have been closed
-            }
-        };
-        return await refreshUserGroups(userId, onProgress);
-    });
-
-    // Posts
-    ipcMain.handle('posts:get-all', async (_, { includeDeleted, status }) => {
-        return await getPosts(includeDeleted, status);
-    });
-
-    ipcMain.handle('posts:create', async (_, postData) => {
-        return await addPost(postData);
-    });
-
-    ipcMain.handle('posts:delete', async (_, { id, force }) => {
-        return await deletePost(id, force);
-    });
-
-    // Updater
-    ipcMain.handle('updater:check', async (_, { channel } = {}) => {
-        const settings = await getUpdateSettings();
-        const ch = channel || settings.channel || 'stable';
-        return await checkForUpdates(ch);
-    });
-
-    ipcMain.handle('updater:get-settings', async () => {
-        return await getUpdateSettings();
-    });
-
-    ipcMain.handle('updater:save-settings', async (_, settings) => {
-        return await saveUpdateSettings(settings);
-    });
-
-    ipcMain.handle('updater:open-download', async (_, { url }) => {
-        openDownloadPage(url);
-        return true;
-    });
-
-    ipcMain.handle('updater:download-update', async () => {
-        return await downloadUpdate();
-    });
-
-    ipcMain.handle('updater:install-update', () => {
-        installUpdate();
-    });
-}
