--- main/preload.js
@@ -1,20 +1,0 @@
-import { contextBridge, ipcRenderer } from 'electron'
-
-const handler = {
-  send(channel, value) {
-    ipcRenderer.send(channel, value)
-  },
-  on(channel, callback) {
-    const subscription = (_event, ...args) => callback(...args)
-    ipcRenderer.on(channel, subscription)
-
-    return () => {
-      ipcRenderer.removeListener(channel, subscription)
-    }
-  },
-  invoke(channel, ...args) {
-    return ipcRenderer.invoke(channel, ...args)
-  },
-}
-
-contextBridge.exposeInMainWorld('ipc', handler)
