--- main/updater.js
@@ -1,227 +1,0 @@
-import { app, shell } from 'electron';
-import { autoUpdater } from 'electron-updater';
-import { readJson, writeJson } from './storage.js';
-
-const GITHUB_OWNER = 'takacore';
-const GITHUB_REPO = 'vrchat-group-scheduler';
-const RELEASES_API_URL = `https://api.github.com/repos/${GITHUB_OWNER}/${GITHUB_REPO}/releases`;
-const SETTINGS_FILE = 'update-settings.json';
-
-// Dynamic import for node-fetch (ESM)
-let fetch;
-async function getFetch() {
-    if (!fetch) {
-        const mod = await import('node-fetch');
-        fetch = mod.default;
-    }
-    return fetch;
-}
-
-/**
- * Get current app version from Electron app module
- */
-function getCurrentVersion() {
-    return app.getVersion();
-}
-
-/**
- * Parse semver string into comparable parts
- */
-function parseSemver(version) {
-    const cleaned = version.replace(/^v/, '');
-    const [main, prerelease] = cleaned.split('-');
-    const [major, minor, patch] = main.split('.').map(Number);
-    return { major, minor, patch, prerelease: prerelease || null };
-}
-
-/**
- * Compare two semver versions
- */
-function compareSemver(a, b) {
-    const va = parseSemver(a);
-    const vb = parseSemver(b);
-
-    if (va.major !== vb.major) return va.major > vb.major ? 1 : -1;
-    if (va.minor !== vb.minor) return va.minor > vb.minor ? 1 : -1;
-    if (va.patch !== vb.patch) return va.patch > vb.patch ? 1 : -1;
-
-    if (!va.prerelease && vb.prerelease) return 1;
-    if (va.prerelease && !vb.prerelease) return -1;
-
-    if (va.prerelease && vb.prerelease) {
-        return va.prerelease.localeCompare(vb.prerelease);
-    }
-
-    return 0;
-}
-
-/**
- * Get update settings from storage
- */
-export async function getUpdateSettings() {
-    const defaults = {
-        channel: 'stable',
-        autoCheck: true
-    };
-    const settings = await readJson(SETTINGS_FILE, defaults);
-    return { ...defaults, ...settings };
-}
-
-/**
- * Save update settings to storage
- */
-export async function saveUpdateSettings(settings) {
-    const current = await getUpdateSettings();
-    const merged = { ...current, ...settings };
-    await writeJson(SETTINGS_FILE, merged);
-    return merged;
-}
-
-/**
- * Check for updates from GitHub Releases (for channel-aware checking)
- */
-export async function checkForUpdates(channel) {
-    const fetchFn = await getFetch();
-    const currentVersion = getCurrentVersion();
-
-    try {
-        const response = await fetchFn(RELEASES_API_URL, {
-            headers: {
-                'Accept': 'application/vnd.github.v3+json',
-                'User-Agent': `VRChatGroupScheduler/${currentVersion}`
-            }
-        });
-
-        if (!response.ok) {
-            throw new Error(`GitHub API error: ${response.status} ${response.statusText}`);
-        }
-
-        const releases = await response.json();
-
-        if (!releases || releases.length === 0) {
-            return {
-                updateAvailable: false,
-                currentVersion,
-                latestVersion: currentVersion,
-                releaseNotes: '',
-                downloadUrl: '',
-                isBeta: false
-            };
-        }
-
-        let targetRelease;
-        if (channel === 'beta') {
-            targetRelease = releases[0];
-        } else {
-            targetRelease = releases.find(r => !r.prerelease);
-        }
-
-        if (!targetRelease) {
-            return {
-                updateAvailable: false,
-                currentVersion,
-                latestVersion: currentVersion,
-                releaseNotes: '',
-                downloadUrl: '',
-                isBeta: false
-            };
-        }
-
-        const latestVersion = targetRelease.tag_name.replace(/^v/, '');
-        const updateAvailable = compareSemver(latestVersion, currentVersion) > 0;
-
-        return {
-            updateAvailable,
-            currentVersion,
-            latestVersion,
-            releaseNotes: targetRelease.body || '',
-            downloadUrl: targetRelease.html_url,
-            isBeta: targetRelease.prerelease || false,
-            publishedAt: targetRelease.published_at
-        };
-    } catch (error) {
-        console.error('Update check failed:', error);
-        throw error;
-    }
-}
-
-// --- electron-updater Auto Update ---
-
-/**
- * Configure and initialize electron-updater autoUpdater
- */
-export function initAutoUpdater(mainWindow) {
-    autoUpdater.autoDownload = false;
-    autoUpdater.autoInstallOnAppQuit = true;
-
-    autoUpdater.on('update-available', (info) => {
-        console.log('[AutoUpdater] Update available:', info.version);
-        mainWindow.webContents.send('updater:auto-update-available', {
-            version: info.version,
-            releaseNotes: info.releaseNotes || '',
-        });
-    });
-
-    autoUpdater.on('download-progress', (progress) => {
-        mainWindow.webContents.send('updater:download-progress', {
-            percent: Math.round(progress.percent),
-            bytesPerSecond: progress.bytesPerSecond,
-            transferred: progress.transferred,
-            total: progress.total,
-        });
-    });
-
-    autoUpdater.on('update-downloaded', (info) => {
-        console.log('[AutoUpdater] Update downloaded:', info.version);
-        mainWindow.webContents.send('updater:update-downloaded', {
-            version: info.version,
-        });
-    });
-
-    autoUpdater.on('error', (err) => {
-        console.error('[AutoUpdater] Error:', err);
-        mainWindow.webContents.send('updater:error', {
-            message: err.message || 'アップデートエラーが発生しました',
-        });
-    });
-}
-
-/**
- * Check for updates using electron-updater
- */
-export async function autoCheckForUpdates(channel) {
-    autoUpdater.allowPrerelease = channel === 'beta';
-    try {
-        await autoUpdater.checkForUpdates();
-    } catch (err) {
-        console.error('[AutoUpdater] Check error:', err);
-    }
-}
-
-/**
- * Download the update
- */
-export async function downloadUpdate() {
-    try {
-        await autoUpdater.downloadUpdate();
-    } catch (err) {
-        console.error('[AutoUpdater] Download error:', err);
-        throw err;
-    }
-}
-
-/**
- * Install the update and restart the app
- */
-export function installUpdate() {
-    autoUpdater.quitAndInstall(false, true);
-}
-
-/**
- * Open the download URL in the system browser (fallback for macOS)
- */
-export function openDownloadPage(url) {
-    if (url) {
-        shell.openExternal(url);
-    }
-}
