+++ next.config.js
@@ -1,12 +1,9 @@
 /** @type {import('next').NextConfig} */
 module.exports = {
   output: 'export',
-  distDir: process.env.NODE_ENV === 'production' ? '../app' : '.next',
+  assetPrefix: './',
   trailingSlash: true,
   images: {
     unoptimized: true,
-  },
-  webpack: (config) => {
-    return config
   },
 }
