+++ package.json
@@ -1,32 +1,15 @@
 {
   "private": true,
   "name": "vrchat-group-notify-scheduler",
-  "description": "VRChat Group Notify Scheduler",
-  "version": "1.1.5",
+  "description": "VRChat Group Notify Scheduler Extension",
+  "version": "2.0.0",
   "author": "TakaAizu",
-  "main": "app/background.js",
   "scripts": {
-    "dev": "nextron",
-    "build": "nextron build",
-    "build:mac": "nextron build --mac",
-    "build:win": "nextron build --win --x64",
-    "build:all": "nextron build --mac --win --x64",
-    "postinstall": "electron-builder install-app-deps"
+    "dev": "next dev",
+    "build": "next build && node fix-extension.js"
   },
   "dependencies": {
-    "cookie": "^1.1.1",
-    "electron-log": "^5.4.3",
-    "electron-serve": "^1.3.0",
-    "electron-store": "^8.2.0",
-    "electron-updater": "^6.8.3",
-    "node-fetch": "^3.3.2",
-    "node-schedule": "^2.1.1"
-  },
-  "devDependencies": {
-    "electron": "^34.0.0",
-    "electron-builder": "^24.13.3",
     "next": "^14.2.4",
-    "nextron": "^9.5.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   }
