+++ public/background/storage.js
@@ -1,0 +1,18 @@
+// background/storage.js
+export const storage = {
+    async get(keys) {
+        return new Promise((resolve) => {
+            chrome.storage.local.get(keys, (result) => {
+                resolve(result);
+            });
+        });
+    },
+
+    async set(items) {
+        return new Promise((resolve) => {
+            chrome.storage.local.set(items, () => {
+                resolve();
+            });
+        });
+    }
+};
