+++ public/manifest.json
@@ -1,0 +1,30 @@
+{
+    "manifest_version": 3,
+    "name": "VRChat Group Scheduler",
+    "version": "2.0.0",
+    "description": "Schedule and automate group announcements on VRChat.",
+    "permissions": [
+        "storage",
+        "alarms",
+        "cookies",
+        "notifications",
+        "scripting"
+    ],
+    "host_permissions": [
+        "https://*/*",
+        "http://*/*"
+    ],
+    "background": {
+        "service_worker": "background/background.js",
+        "type": "module"
+    },
+    "action": {
+        "default_popup": "index.html",
+        "default_title": "VRChat Group Scheduler"
+    },
+    "icons": {
+        "16": "icons/icon16.png",
+        "48": "icons/icon48.png",
+        "128": "icons/icon128.png"
+    }
+}
