--- renderer/pages/index.jsx
@@ -1,12 +1,0 @@
-import { useEffect } from 'react';
-import { useRouter } from 'next/router';
-
-export default function Index() {
-    const router = useRouter();
-
-    useEffect(() => {
-        router.push('/home');
-    }, []);
-
-    return null;
-}
